djangoProjectdw/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-autocomplete-light==3.9.5
django-awesomplete==0.5.2
django-chartjs==2.3.0
django-filter==23.2
django-import-export==3.2.0
django-pdfkit==0.3.1
django-simple-autocomplete==1.11
django-tags-input==6.0.0
docxcompose==1.4.0
docxtpl==0.16.6
et-xmlfile==1.1.0
fonttools==4.39.3
html5lib==1.1
item==0.0.3
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2

odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.0

pdfkit==0.6.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pyHanko==0.18.0
pyhanko-certvalidator==0.22.0
pyparsing==3.0.9
pypdf==3.8.1
python-docx==0.8.11

pytz-deprecation-shim==0.1.0.post0


requests==2.29.0

sqlparse==0.4.3
svglib==1.5.1

weblib==0.1.30
xhtml2pdf==0.2.10

zipp==3.15.0
gunicorn==20.1.0
